+++ client/src/utils/date.ts
@@ -1,3 +1,12 @@
+import {
+  getDate,
+  endOfMonth,
+  getDay,
+  startOfMonth,
+  getWeeksInMonth,
+  eachDayOfInterval,
+} from './date-fns'
+
 const A_MINUTE = 1000 * 60
 const AN_HOUR = 1000 * 60 * 60
 const A_SECOND = 1000
@@ -54,4 +63,26 @@ export function timerFormat(seconds: number) {
   return formattedHour === '00'
     ? `${formattedMinute}:${formattedSecond}`
     : `${formattedHour}:${formattedMinute}:${formattedSecond}`
+}
+
+export function getWeekRows(date: Date): (0 | Date)[][] {
+  const lastDate = getDate(endOfMonth(date))
+  const startBlankCount = getDay(startOfMonth(date))
+  const endBlankCount = 7 - ((startBlankCount + lastDate) % 7)
+  const weekCount = getWeeksInMonth(date)
+
+  const allDates = [
+    ...Array(startBlankCount).fill(0),
+    ...eachDayOfInterval({
+      start: startOfMonth(date),
+      end: endOfMonth(date),
+    }),
+    ...Array(endBlankCount).fill(0),
+  ]
+
+  const rows = Array(weekCount)
+    .fill(0)
+    .map((_, i) => [...allDates].splice(i * 7, 7))
+
+  return rows
 }
